refact : mapcomponent data layer 분리
